Added pre- and postinstall scripts to the package.
The `preinstall' script will stop a previous nginx installation if it is
running.  The `postinstall' script will test the configuration and start
nginx.

Added the necessary makefile variable, dependency and `pkgbuild' command
line option to include the scripts.

Also changed the uninstall script to include the "set -euo pipefail"
directive.