Inserted a version footer to nginx index.html page.
Added a build step to insert a small footer into the nginx index.html
page that shows the nginx version and installer package
version/revision, with a link the the GitHub repository.

Also improved the makefile by instituting a variable naming convention:
variables representing files and directories in the final $(TMP)/pkg
directory are prefixed with `pkg_'.