Build universal (fat) binaries for both arm64 and x86_64.
Add `-arch` flags to compilation and linking commands to produce a
universal binary that supports both Intel and Apple Silicon Macs.

Created an `archs` variable at the top of the makefile to control the
list of architectures, and a _compilation flags_ section that builds the
CLANG `-archs` flags and adds them to `CFLAGS`.  Also defined a `LINK`
variable, since the nginx makefile doesn't use `LDFLAGS`, but defines
the linker command as `LINK`.

Explicitly pass `CFLAGS` and `LINK` to the nginx `make` and `make
install` invocations.

Thanks to [this post by Chris
Cannam](https://thebreakfastpost.com/2021/01/05/on-macos-arm64-and-universal-binaries/)
for help in finding the correct flags.