Statically link zlib to nginx.
Pull the zlib 1.2.11 source into the project and configure nginx to
build using the included zlib source.  Added make targets to copy the
zlib source distribution into `$(TMP)` to keep the project source
directory clean.

Added a test to `make check` to validate that the nginx executable
doesn't link to any non-system dylibs.  Added the `tools/` directory and
the `dylibs` command that wraps `otool` to get the dylibs list from the
nginx executable.

Updated the build report, the installer welcome page and the default
HTML footer to include the zlib version info.