Fix dependency graph problems when making default target.
The `signed-package` default target was always out of date for three
reasons:

    - signing the `nginx` binary modified it in place, making the
      `uninstall-nginx` target out of date
    - when nginx builds pcre, it modifies the pcre `configure` file
      during its build, making the `configured.stamp.txt` file (marking
      nginx's configuration) out of date
    - similary, when building zlib, the zlib `Makefile` and `zconf.h`
      files are modified, making `configured.stamp.txt` out of date

Fixed the first problem by copying the `nginx` binary from the
`nginx/install` directory to the `pkg` directory separately from the
rest of the installed nginx files, and making the action for the `nginx`
binary both copy and sign.

Made an initial quick fix for the other two problems by making the temp
copy of the pcre and zlib distribution files "order-only" prerequisites
to the `configured.stamp.txt` target, so that the modification times of
these files aren't considered.