A better dependency graph for the nginx configuration target.
Change the `configured.stamp.txt` target (representing running nginx's
configure script) so that it depends directly on the pcre and zlib
distribution sources and omit targets for the copy of those sources into
`$(TMP)`.  This works around a problem where the nginx build modifies a
few of the pcre and zlib sources in-place, causing the nginx configure
target to look out of date.

Change the action for that target to completely remove the temp copies
of pcre and zlib and re-copy them over before running nginx's configure
script.  This should cause nginx to be re-configured and rebuilt if
there is any change to pcre or zlib (e.g. when upgrading versions).