Fix problem with `notarize` dependency graph.
Because stamping modified the unnotarized package in place, it made the
`stapled.stamp.txt` file always seem out of date, since the dependency
graph points back to the package.

Change the `signed-package` default target so that it depends on
`nginx-<version>-unnotarized.pkg` in the temp directory, and change the
`notarize` target so that it depends on `nginx-<version>.pkg` in the
project root.

Changed intermediate targets to match, and changed the final target that
runs the `stapler` command to first copy
`nginx-<version>-unnotarized.pkg` into place _then_ staple it.

Also added a line break to the installer welcome page.